Test Griddler::Controller with an anonymous controller
